feat: improve functionality in deck
Add if statements in the click card event listener. Checks
if the target of click is the <li> element and if the openedCards array
has been emptied by the setTimeout(). Also, checks if the clicked
card is the first opened or not.

These changes help solve previous problems in functionality. Game logic
crashed when the player clicked more that 2 cards, or when the player
clicked on already matched cards, or even when he clicked the same opened
card.

Problems now resolved.